Use specific role-based groups for authentication
This removes the need for the general "SU role", since we use the specific
roles (veileder, saksbehandler, attestant) more actively. We could probably now
remove the general role from Vault.

Do note that we don't have proper authorization on routes in place yet at this
point. Implementing this should be the next step.