Beregning: distribute all fradrag evenly over the whole period
The behaviour for what to do if one or more months in a stønadsperiode have
exceptionally high fradrag is not defined.
For example:
Stønadsperiode 1. january through 31. december, and one fradrag in march having
a beløp of 1000000 NOK. Without this workaround the fradrag would only be
calculated as part of the månedsberegning for march, and since fradrag are
limited upwards to the satsbeløp, we would "discard" most of that million. The
søker would then get more money than they probably should.

Exactly how this scenario _should_ work is at the moment not defined, hence the
need for this workaround to at least have the system behave as previously (all
fradrag are at least part of the calculation).

Co-authored-by: Jacob Meidell <jacob.meidell@nav.no>